E3_MettreDispositionIA/collab_tool/requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Pillow==10.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
streamlit==1.31.1 